pyproject.toml
[tool.poetry]
name = "apart-eval-hackathon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "apart_eval_hackathon"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
nltk = "^3.8.1"
datasets = "^2.14.4"
cachier = "^2.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
